Simplify the public API
